pyproject.toml
[project]
name = "theatris-rpo"
version = "0.1.0"
description = "Remotely OSC-controlled video-player for raspberry pi"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.4.0",
    "pygobject>=3.54.5",
    "pygobject-stubs>=2.16.0",
    "pykms",
    "python-osc>=1.9.3",
    "python-oscquery>=0.3.0",
    "returns>=0.26.0",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.uv.sources]
pykms = { git = "https://github.com/tomba/pykms" }

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = ["--import-mode=importlib", ]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyfakefs>=6.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.uv.build-backend]
module-name = "theatris_rpo"
